refactor(Playwright): Adapted enhance-resolve Jest 27 workaround
While debugging a bug in tsconfig.json we saw, that the failure
messages did not show important details. Slightly adapted
the enhance-resolve workaround for Jest 27 and especially added some
logging on failure to ease debugging.
