+++ itest/empty.js
@@ -1,0 +1,1 @@
+// Used when a browser field is set to false
