--- itest/jest-9771-workaround-empty.js
@@ -1,1 +1,0 @@
-// used in jest-9771-resolve-workaround.js
